zadanie17

+++ zadanie17/TestLib/Models/Figure.cs
@@ -8,7 +8,7 @@ namespace Figurelib
 
         }
 
-        public virtual int Perim(int p)
+        public virtual double Perim(double p)
         {
             return p;
         }
